// src/api/model/test.ts
/**
 * Generated by orval v6.26.0 🍺
 * Do not edit manually.
 * student-hub-api
 * Api Documentation for StudentHub API
 * OpenAPI spec version: 0.0
 */
import type { TestSubTopicsItem } from './testSubTopicsItem';

export interface Test {
  bestGrade?: number;
  bestSubTopic?: string;
  date?: string;
  subTopics?: TestSubTopicsItem[];
  testEventId?: number;
  title?: string;
  userId?: number;
  worstSubTopic?: string;
}